fa.asreml {ASExtras4}	R Documentation

Factor analytic variance models.

Description

Summarises an asreml model in which factor analytic (FA) variance structures have been fitted. The function aids in interpreting FA models fitted to genotype by environment effects in the analysis of plant breeding multi-environment trial (MET) data.

Usage

fa.asreml(
  object,
  uniplot = TRUE,
  uniplot.tol = 0.85,
  uniplot.cex = 0.5,
  trunc.char = c(1, 6),
  blups = TRUE,
  regplot = TRUE,
  addedplot = TRUE,
  g.list = NULL,
  heatmap = TRUE,
  heatmap.ord = "asis",
  agnes.method = "average"
)

Arguments

object
An asreml object.

uniplot
if TRUE (the default), pairwise plots of FA loadings are produced for k > 1.

uniplot.tol
A numeric value between 0 and 1 that affects the display of pairs of loadings in the uniplots. Those pairs with a (scaled) euclidean distance greater than uniplot.tol are drawn with solid blue lines, otherwise they appear as red dashed lines.

uniplot.cex
Character expansion (relative) size for plotting (the default is 0.5).

trunc.char
A numeric vector of length 2 specifying the truncation of loading labels for use in plotting. The default (c(1,6)) is to use the first 6 characters of the environment factor levels, otherwise if NULL the complete level names are used.

blups
If TRUE (the default), summaries of estimates from the asreml object are produced. These include BLUPs of genotype by environment effects and genotype scores, and REML estimates of FA variance parameters.

regplot
If TRUE (the default), regression plots of genotype by environment effects against FA loadings are produced.

addedplot
If TRUE (the default), added variable plots of genotype by environment effects against FA loadings are produced.

g.list
A character vector comprising the subset of genotype levels to be used in the regression and added variable plots. The default (NULL) is to use the complete set of genotype levels.

heatmap
If TRUE (the default), a heatmap of the estimated FA correlation matrix is produced.

heatmap.ord
A character string ("asis", "cluster") or vector of environment levels in the order desired for plotting the heatmap. The default is "asis", where the rows and columns of the correlation matrix are maintained in environment level order. The "cluster" option invokes the agnes clustering algorithm on the FA correlation matrix then orders the rows and columns accordingly.

agnes.method
Clustering strategy; the default is "average".

Details

In the analysis of MET data, the FA model is applied to the vector of random genotype by environment effects (Smith et al, 2001). The model can be thought of as a (multiple) regression of genotype by environment effects on environment covariates (also called loadings) with separate slopes (scores) for each genotype. We denote the number of terms or factors in the model by k, the number of environments by t and varieties by m.

The variance parameters associated with this FA model are the loadings (t_k of these) and the so-called specific variances (t of these) which can be regarded as the regression lack-of-fit or residual variances for individual environments. Together the loadings and specific variances provide a form for the genetic variance matrix for environments (diagonal elements are the genetic variances for individual environments; off-diagonal elements are the genetic covariances between pairs of environments). fa.asreml returns the REML estimate of the genetic variance matrix (additionally converted to a correlation matrix). A graphical depiction of the correlation matrix as a heatmap is optionally available. If requested the rows and columns on the heatmap may be re-ordered on the basis of the application of the agnes clustering algorithm to the genetic correlation matrix (See Cullis et al, 2010).

fa.asreml also returns REML estimates of the loadings and specific variances. The loadings have been rotated to a principal component solution to aid interpretation. The loadings may be depicted graphically as uniplots in which pairs of factors are plotted against each other. The pair of loadings for an individual environment is represented as a line from the origin. To assist interpretation, the loadings are first converted to the correlation rather than covariance scale. The unit circle is imposed on each uniplot as it reflects the upper limit where 100% of genetic variance is explained by the two factors being plotted (See Cullis et al, 2010).

fa.asreml also returns BLUPs of the genotype effects for each environment (t_m of these) and BLUPs of the genotype scores for each factor (k_m of these).

Plots that reflect the multiple regression implicit in the FA model are also optionally available. These may either be regression or added variable plots. With the former, the genotype by environment BLUPs are plotted against the loadings, with a separate panel for each factor and each (pre-specified) genotype. The BLUPs are plotted as points. Super-imposed on each panel is the regression line, that is, with slope given by the genotype score for that factor. The added variable plots are similar except that the points on the panels for factor i, i = 2 ... k are the regression residuals associated with the first i-1 factors.

Note that several FA models are allowed within the asreml call. For example if pedigree information is included in the analysis there may be separate FA models (possibly of different order) for the additive and non-additive genotype by environment effects (Cullis et al, 2010). We let f denote the number of FA terms in the model (typically f <= 2). Also, the factor names Environment and Genotype are used generically here, however, fa.asreml uses the factor names appearing in the FA term(s) in the asreml model.

Value

A list with the following components:

blups

a list of length f. Each element is itself a list, with elements:

blups

A data-frame of length m_t containing the BLUPs of the genotype effects for each environment and the associated regression BLUPs (predictions from the FA regression terms). These two differ in that the latter does not include the regression residuals. Also given is the number of non-missing y-values (may be zero) for each genotype in each environment.

scores

A data-frame of length m_k containing the BLUPs of the genotype scores for each factor.

gammas

A list of length f where each element is a list with components:

Gmat

The REML estimate of the genetic variance matrix (dimension t \times t)

Cmat

The REML estimate of the genetic correlation matrix (dimension t \times t)

rotated loads

A matrix of dimension t \times k containing the REML estimates of the FA loadings, rotated to a principal component solution.

rotated loads - c

A matrix of dimension t \times k containing the REML estimates of the FA loadings, rotated to a principal component solution and re-scaled to align with Cmat rather than Gmat.

specific var

A vector of length t containing the REML estimates of the specific variances.

site %vaf

A matrix of dimension t \times k containing, for each environment, the percentage of genetic variance accounted for by the FA regression, both as a whole and by individual factors.

total %vaf

A scalar giving the total percentage of genetic variance accounted for by the FA regression.

uniplots

A list of length f. Each element may be a trellis object containing a pairwise plot of factor loadings 1 and 2 (k=2), or a list of k(k-1)/2 trellis objects (k>2), one element for each pairwise plot.

regplots

A list of length f. Each element is a trellis object containing plots of the genotype by environment BLUPs against factor loadings. There is a separate panel for each factor and genotype specified in g.list. Superimposed on each panel is the associated regression line with slope given by the genotype score for that factor.

addedplots

A list of length f. Each element is a trellis object similar to that in regplots except it contains added variable plots. The points on the panels for factor i, i = 2 ... k are the regression residuals associated with the first i-1 factors. When k=1 the graphics in regplots and addedplots are identical.

heatmap

A list of length f. Each element is a trellis plot containing a heatmap of the genetic correlation matrix. The rows and columns will be ordered as per the heatmap.ord argument.

agnes

A list of length f where each element is the standard output from the agnes function. Returned only if the heatmap.ord argument was specified with the "cluster" option.

References

Cullis BR, Smith AB, Beek C, Cowling WA (2010). “Analysis of Yield and Oil from a Series of Canola Breeding Trials. Part II: Exploring VxE using Factor Analysis.” Genome, 53, 1002-1016. Smith AB, Cullis BR, Thompson R (2001). “Analyzing Variety by Environment Data Using Multiplicative Mixed Models and Adjustments for Spatial Field Trend.” Biometrics, 57, 1138-1147.

[Package ASExtras4 version 1.0 Index]