faces.asreml {ASExtras4}	R Documentation

Variogram faces

Description

Computes and plots the approximate coverage intervals for the faces of the semi-variogram of the residuals from a spatially defined experiment.

Usage

faces.asreml(
  obj,
  sim = 100,
  trace = FALSE,
  trim = 20,
  main = "Approximate 95% coverage interval"
)

Arguments

obj
An asreml object.

sim
The number of simulations, default is 100.

trace
If TRUE, asreml convergence montioring is written to the R console (the default is FALSE).

trim
Variogram values based on less than trim points are discarded (the default is 20).

main
Title for the convergence plots.

Details

For each of sim iterations, a new dataframe is constructed from sampling the distributions of the random model terms. Only single section models with simple variance parameters in the random formula are allowed.

Value

A list with the following components:

plots

A list of length two containing the trellis plots for the variogram faces.

vparameters

A two-column matrix of the fitted and (mean) simulated random components.

tau

A two-column matrix of the fitted and (mean) simulated fixed effects.

variogram

The data frame of variogram summaries used in plots.

References

Stefanova KT, Smith AB, Cullis BR (2009). “Enhanced Diagnostics for the Spatial Analysis of Field Trials.” Journal of Agricultural, Biological, and Environmental Statistics, 14(4), 392-410.

[Package ASExtras4 version 1.0 Index]