met.asreml {ASExtras4}	R Documentation

Multi-environment trials.

Description

Plot the sample variogram and spatial trends in residuals from the meta-analysis of a series of field experiments.

Usage

met.asreml(
  object,
  plottype = "together",
  type = "default",
  pages = NA,
  trim = 6,
  spatial = "plot",
  npanels = NA,
  swap.axes = TRUE,
  new.device = FALSE,
  display = TRUE,
  ...
)

Arguments

object
An object of class asreml, or class met.asreml from a previous call.

plottype
A character string ("together", "separate") specifying the layout of the plots. The default "together" draws a trellis display of residuals against the spatial dimensions, and a variogram of the residuals in the one graphics frame. Otherwise, the plots appear in separate frames.

type
Type of residuals, see residuals.asreml.

pages
A numeric or character vector identifying the pages to display from a previous call to met.asreml. If NA (the default), all plots are (re)displayed.

trim
An integer specifying the minimum number of observations for a point to be included in the semi-variogram; the default is 6.

spatial
Whether to include a nugget effect; see residuals.asreml.

npanels
If plottype == "separate", npanels specifies the number of trellis panels per page. If NA (the default), a suitable default is calculated.

swap.axes
If TRUE (the default), the dimension with the highest number of levels is on the x-axis, otherwise the order specified in the residual formula is retained.

new.device
if TRUE (default is FALSE), a new trellis device is openned for each plot. Ignored if plottype == "separate".

display
If TRUE (the default), render the plots on the default display. met.asreml silently returns a list of the trellis objects produced.

...
Additional arguments to be passed to the lattice functions.

Details

met.asreml displays the residuals from a series of one or more experiments in a set of "pairs" of trellis plots. Each pair contains a trellis plot of the residuals against the factors defining the dimensions in the residual formula of an asreml model, and an empirical semi-variogram of the residuals against these factors. Each pair may appear in a single trellis frame (plottype="together"), or in separate trellis devices.

Value

An object of class met.asreml containing a list of trellis objects, with elements named by the levels of the factor defining the sections in the residual model formula.

[Package ASExtras4 version 1.0 Index]